refactor(api): move stage-acquisition-repository to evaluation scope
