+++ api/lib/domain/models/index.js
@@ -117,7 +117,6 @@ import { SessionPublicationBatchResult } from './SessionPublicationBatchResult.j
 import { ShareableCertificate } from './ShareableCertificate.js';
 import { Skill } from './Skill.js';
 import { Solution } from './Solution.js';
-import { StageAcquisition } from '../../../src/evaluation/domain/models/StageAcquisition.js';
 import { Student } from './Student.js';
 import { Tag } from './Tag.js';
 import { TargetProfile } from './TargetProfile.js';
