+++ api/lib/domain/usecases/get-user-campaign-assessment-result.js
@@ -3,7 +3,7 @@
 import { NoCampaignParticipationForUserAndCampaign, NotFoundError } from '../errors.js';
 import * as defaultCompareStageAndAcquiredStagesService from '../../../src/evaluation/domain/services/stages/stage-and-stage-acquisition-comparison-service.js';
 import * as defaultStageRepository from '../../../src/evaluation/infrastructure/repositories/stage-repository.js';
-import * as defaultStageAcquisitionRepository from '../../infrastructure/repositories/stage-acquisition-repository.js';
+import * as defaultStageAcquisitionRepository from '../../../src/evaluation/infrastructure/repositories/stage-acquisition-repository.js';
 import * as defaultParticipantResultRepository from '../../infrastructure/repositories/participant-result-repository.js';
 
 const getUserCampaignAssessmentResult = async function ({
