+++ api/lib/domain/usecases/stages/handle-stage-acquisition.js
@@ -3,7 +3,7 @@
 import * as defaultSkillRepository from '../../../infrastructure/repositories/skill-repository.js';
 import * as defaultStageRepository from '../../../../src/evaluation/infrastructure/repositories/stage-repository.js';
 import * as defaultCampaignRepository from '../../../infrastructure/repositories/campaign-repository.js';
-import * as defaultStageAcquisitionRepository from '../../../infrastructure/repositories/stage-acquisition-repository.js';
+import * as defaultStageAcquisitionRepository from '../../../../src/evaluation/infrastructure/repositories/stage-acquisition-repository.js';
 import * as defaultCampaignParticipationRepository from '../../../infrastructure/repositories/campaign-participation-repository.js';
 import * as defaultKnowledgeElementRepositoryRepository from '../../../infrastructure/repositories/knowledge-element-repository.js';
