+++ api/src/evaluation/domain/usecases/index.js
@@ -7,10 +7,10 @@ import * as answerRepository from '../../infrastructure/repositories/answer-repo
 import * as assessmentRepository from '../../../shared/infrastructure/repositories/assessment-repository.js';
 import * as competenceRepository from '../../../shared/infrastructure/repositories/competence-repository.js';
 import * as competenceEvaluationRepository from '../../infrastructure/repositories/competence-evaluation-repository.js';
+import * as stageAcquisitionRepository from '../../infrastructure/repositories/stage-acquisition-repository.js';
 import * as stageCollectionForTargetProfileRepository from '../../infrastructure/repositories/stage-collection-repository.js';
 import * as stageRepository from '../../infrastructure/repositories/stage-repository.js';
 import * as feedbackRepository from '../../infrastructure/repositories/feedback-repository.js';
-import * as stageRepository from '../../../../lib/infrastructure/repositories/stage-repository.js';
 import * as targetProfileForAdminRepository from '../../../../lib/infrastructure/repositories/target-profile-for-admin-repository.js';
 import { getCompetenceLevel } from '../services/get-competence-level.js';
 
@@ -26,6 +26,7 @@ const dependencies = {
   competenceEvaluationRepository,
   competenceRepository,
   stageCollectionForTargetProfileRepository,
+  stageAcquisitionRepository,
   feedbackRepository,
   stageRepository,
   targetProfileForAdminRepository,
